Switch Python scripts to Python 3
The scripts had already the majority of the work needed for Python 3,
and Python 2 is EOL since Jan 1st 2020, so just the scripts to Python 3.

The last change required is to fix the repr() on chars, due to the text
vs data changes in Python 3: repr() prints more Unicode representations
now, breaking the C output (e.g. \r now is lterally printed as such,
removing all the content in the line before that). As fix, use the
workaround proposed in PEP-3138 [1]. This fixes the generate_fsm_test_sh
test.

[1] https://www.python.org/dev/peps/pep-3138/#id10